Create PartScriptableObject.cs
Contains the gameobject, an ID (which is just any number rn), and the connecting directions. These are directions that the part can connect to the ship from. for example, the engine can only connect from the top side of itself, as the bottom is ejecting flames and stuff, so the only direction is "above". Other directions are "below", "left", and "right".