Fixed some logic involving guard clauses for ShipBuildingGrid.cs
Moved the check to see if a tile is empty to be inside the CanPlacePart() method, as that check should always be done when seeing if a part can be placed. This fixes a bug that made parts dragged to another tile able to be placed on the same tile.